+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/appearance/AppearanceModule.kt
@@ -1,16 +1,16 @@
-package io.horizontalsystems.bankwallet.modules.settings.launch
+package io.horizontalsystems.bankwallet.modules.settings.appearance
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 
-object LaunchPageModule {
+object AppearanceModule {
 
     class Factory() : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val service = LaunchPageService(App.localStorage)
-            return LaunchPageViewModel(service) as T
+            val service = AppearanceService(App.localStorage)
+            return AppearanceViewModel(service) as T
         }
     }
